refactor: 收敛附件生命周期与模型解析
