+++ backend/package/yuxi/services/thread_files_service.py
@@ -16,7 +16,8 @@ from yuxi.agents.backends.sandbox import (
     virtual_path_for_thread_file,
 )
 from yuxi.repositories.conversation_repository import ConversationRepository
-from yuxi.services.conversation_service import materialize_attachment_records, require_user_conversation
+from yuxi.services.attachment_service import materialize_attachment_records
+from yuxi.services.conversation_service import require_user_conversation
 from yuxi.services.mention_search_service import invalidate_mention_cache, invalidate_workspace_mention_cache
 from yuxi.utils.datetime_utils import utc_isoformat_from_timestamp
 from yuxi.utils.paths import VIRTUAL_PATH_PREFIX
