+++ backend/test/unit/services/test_subagent_run_service.py
@@ -214,7 +214,12 @@ def _patch_run_record_creation(
     async def get_system_options(_option, _db=None):
         return {"default_model": "system-default:model"}
 
-    monkeypatch.setattr(type(service_module.system_options), "get", get_system_options)
+    monkeypatch.setattr(type(agent_run_service.system_options), "get", get_system_options)
+    monkeypatch.setattr(
+        agent_run_service.model_cache,
+        "get_model_info",
+        lambda _spec: SimpleNamespace(model_type="chat"),
+    )
 
     class _FakeContext:
         def __init__(self):
